slide and ground pound code + more fucntions
added most not all of the things that are needed for the slide and ground pound. the ground pound is finished but I need to finish the slide. I need to shrink the hitbox. I made separate functions for the basic movement and the other movements like the ground pound / slide.