camera limit system and updated map
started working on a system that lest you change the limits of the camera on the fly. its still WIP. updated a map so now it has more thigns to do and introdooces more mechanics. It also now has an ending spot or the end of the level also a WIP